Removed orphanremoval annotations
Orphan removal would mean that, if I removed, for example, a commune belonging to a certain province, the province would also be removed as well, which is not the desired behavior.
Think the commune 'Coquimbo' belonging to the province 'Elqui'. If I remove 'Coquimbo', 'Elqui' would be removed as well, which means all communes referencing the province 'Elqui' would now be invalid.